[FIX] 시야 범위 슬라이더 맥스/민값 표기 & int로 변경

+++ Assets/Scripts/Data/LobbyData.cs
@@ -10,8 +10,8 @@ namespace Data
         public int savotageCooltime;
         public int killCooltime;
         public bool isShowKillerInfo;
-        public float innerEyesightValue;
-        public float outerEyesightValue;
+        public int innerEyesightValue;
+        public int outerEyesightValue;
 
         public static LobbyData CreateDefault()
         {
